0.5

+++ f.h
@@ -1,6 +1,6 @@
 #ifndef F_H_
 #define F_H_
 
-void f(unsigned char *xy_matrix, char * z_buffer, int *bitmap_info, int *coordinates, int *angles, unsigned int *temp);
+void f(unsigned char *xy_matrix, char * z_buffer, int *bitmap_info, int *coordinates, int *angles, double *temp);
 
 #endif // F_H_
